:sparkles::recycle: add transfer and refactor double to float

+++ src/main/kotlin/com/zbank/bankaccount/application/data/AccountBalanceData.kt
@@ -2,7 +2,7 @@ package com.zbank.bankaccount.application.data
 
 import com.zbank.bankaccount.domain.model.account.Account
 
-data class AccountBalanceData(val id: Long, val balance: Double) {
+data class AccountBalanceData(val id: Long, val balance: Float) {
     companion object {
 
         fun from(account: Account) = AccountBalanceData(
