+++ src/main/resources/db/migration/V1606282101000__create_account.sql
@@ -2,7 +2,7 @@ CREATE TABLE account(
     id BIGSERIAL PRIMARY KEY,
     name TEXT,
     cpf TEXT,
-    balance DOUBLE PRECISION,
+    balance FLOAT(3),
     updated_at TIMESTAMP WITH TIME ZONE,
     created_at TIMESTAMP WITH TIME ZONE
 );
